# docs/swagger.yaml
basePath: /
definitions:
  entity.ActionClaim:
    properties:
      claim_type:
        $ref: '#/definitions/entity.ActionClaimType'
      created_at:
        type: string
      expires_at:
        type: string
      game_id:
        type: integer
      holder_wallet:
        type: string
      release_reason:
        type: string
      status:
        $ref: '#/definitions/entity.ActionClaimStatus'
    type: object
  entity.ActionClaimStatus:
    enum:
    - active
    - released
    - expired
    type: string
    x-enum-varnames:
    - ActionClaimStatusActive
    - ActionClaimStatusReleased
    - ActionClaimStatusExpired
  entity.ActionClaimType:
    enum:
    - join
    - cancel
    - reveal
    - expired_follow_up
    type: string
    x-enum-varnames:
    - ActionClaimTypeJoin
    - ActionClaimTypeCancel
    - ActionClaimTypeReveal
    - ActionClaimTypeExpiredFollowUp
  entity.ActivityQueueKey:
    enum:
    - joinable
    - my-active
    - reveal-required
    - expired-attention
    - history
    type: string
    x-enum-varnames:
    - ActivityQueueJoinable
    - ActivityQueueMyActive
    - ActivityQueueRevealRequired
    - ActivityQueueExpiredAttention
    - ActivityQueueHistory
  entity.Game:
    properties:
      bet_amount:
        description: nanotons
        type: integer
      complete_tx_hash:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      fee_receiver_address:
        type: string
      game_id:
        type: integer
      highest_bid_allowed:
        type: integer
      init_tx_hash:
        type: string
      join_tx_hash:
        type: string
      joined_at:
        type: string
      lowest_bid_allowed:
        type: integer
      min_referrer_payout_value:
        type: integer
      payout_amount:
        type: integer
      player_one_address:
        type: string
      player_one_choice:
        description: 1=CLOSED, 2=HEADS, 3=TAILS
        type: integer
      player_one_referrer:
        type: string
      player_two_address:
        type: string
      player_two_choice:
        type: integer
      player_two_referrer:
        type: string
      protocol_version:
        type: integer
      referrer_fee_numerator:
        type: integer
      reveal_tx_hash:
        type: string
      revealed_at:
        type: string
      service_fee_numerator:
        type: integer
      status:
        description: 0-4 (see constants above)
        type: integer
      waiting_timeout_seconds:
        type: integer
      winner_address:
        type: string
    type: object
  entity.PlayerActivityGame:
    properties:
      active_claims:
        items:
          $ref: '#/definitions/entity.ActionClaim'
        type: array
      game:
        $ref: '#/definitions/entity.Game'
      is_owned_by_player:
        type: boolean
      latest_activity_at:
        type: string
      next_action:
        $ref: '#/definitions/entity.PlayerActivityNextAction'
      queue_key:
        $ref: '#/definitions/entity.ActivityQueueKey'
      requires_attention:
        type: boolean
    type: object
  entity.PlayerActivityNextAction:
    enum:
    - join
    - resume_join
    - wait_for_join_window
    - resume_cancel
    - wait_for_cancel_resolution
    - wait_for_opponent
    - reveal
    - resume_reveal
    - wait_for_reveal
    - review_result
    - resume_review_result
    - wait_for_review_window
    - browse_history
    type: string
    x-enum-varnames:
    - ActivityNextActionJoin
    - ActivityNextActionResumeJoin
    - ActivityNextActionWaitForJoinWindow
    - ActivityNextActionResumeCancel
    - ActivityNextActionWaitForCancel
    - ActivityNextActionWaitForOpponent
    - ActivityNextActionReveal
    - ActivityNextActionResumeReveal
    - ActivityNextActionWaitForReveal
    - ActivityNextActionReviewResult
    - ActivityNextActionResumeReview
    - ActivityNextActionWaitForReview
    - ActivityNextActionBrowseHistory
  entity.PlayerActivitySummary:
    properties:
      expired_attention_count:
        type: integer
      history_count:
        type: integer
      joinable_count:
        type: integer
      last_activity_at:
        type: string
      my_active_count:
        type: integer
      reveal_required_count:
        type: integer
      wallet_address:
        type: string
    type: object
...
  rest.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
  rest.ExpiredClaimEnvelope:
    properties:
      claim:
        $ref: '#/definitions/rest.ExpiredClaimResponse'
    type: object
  rest.ExpiredClaimRequest:
    properties:
      wallet_address:
        type: string
    required:
    - wallet_address
    type: object
  rest.ExpiredClaimResponse:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      game_id:
        type: integer
      status:
        type: string
      wallet_address:
        type: string
    type: object
  rest.GameActivityQueueResponse:
    properties:
      attention_required:
        type: boolean
      items:
        items:
          $ref: '#/definitions/entity.PlayerActivityGame'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      queue_key:
        $ref: '#/definitions/entity.ActivityQueueKey'
      summary:
        $ref: '#/definitions/entity.PlayerActivitySummary'
      title:
        type: string
      total:
        type: integer
    type: object
  rest.GameActivitySearchResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.PlayerActivityGame'
        type: array
      limit:
        type: integer
      message:
        type: string
      offset:
        type: integer
      query:
        type: string
      queue_scope:
        type: string
      summary:
        $ref: '#/definitions/entity.PlayerActivitySummary'
      total:
        type: integer
    type: object
...
host: localhost:8090
info:
  contact:
    name: POD Game Team
  description: Backend service for TON blockchain gambling game with real-time WebSocket
    updates
  license:
    name: BUSL-1.1
  title: POD Game Backend API
  version: 1.0.0
paths:
...
  /api/v1/games/{gameId}/expired-claim:
    delete:
      consumes:
      - application/json
      description: Releases the current expired-follow-up claim held by the requesting
        wallet.
      parameters:
      - description: Game ID
        in: path
        minimum: 1
        name: gameId
        required: true
        type: integer
      - description: Expired claim release request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.ExpiredClaimRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Claim released
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Release expired-follow-up claim
      tags:
      - games
    get:
      consumes:
      - application/json
      description: Returns the current additive expired-follow-up claim for the specified
        game.
      parameters:
      - description: Game ID
        in: path
        minimum: 1
        name: gameId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.ExpiredClaimEnvelope'
        "204":
          description: No expired-follow-up claim exists
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get expired-follow-up claim
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Creates or resumes an additive expired-follow-up claim for an ended
        game that still requires attention.
      parameters:
      - description: Game ID
        in: path
        minimum: 1
        name: gameId
        required: true
        type: integer
      - description: Expired claim request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.ExpiredClaimRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/rest.ExpiredClaimEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Create expired-follow-up claim
      tags:
      - games
...
  /api/v1/games/activity/{queueKey}:
    get:
      consumes:
      - application/json
      description: Returns queue-oriented activity items for joinable, my-active,
        reveal-required, expired-attention, or history views.
      parameters:
      - description: Queue key
        enum:
        - joinable
        - my-active
        - reveal-required
        - expired-attention
        - history
        in: path
        name: queueKey
        required: true
        type: string
      - description: Optional wallet address used for player-relative queue classification
        in: query
        name: wallet
        type: string
      - default: 20
        description: Maximum number of results
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.GameActivityQueueResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get a player-facing activity queue
      tags:
      - games
  /api/v1/games/activity/search:
    get:
      consumes:
      - application/json
      description: Searches queue-oriented activity items by wallet, opponent, or
        game identifier.
      parameters:
      - description: Search term
        in: query
        name: q
        type: string
      - description: Optional wallet address
        in: query
        name: wallet
        type: string
      - description: Optional queue scope
        in: query
        name: queue
        type: string
      - default: 20
        description: Maximum number of results
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of results to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.GameActivitySearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Search activity surfaces
      tags:
      - games
...
  /api/v1/users/{address}/activity-summary:
    get:
      consumes:
      - application/json
      description: Returns the additive queue counts used by the activity shell for
        the specified wallet.
      parameters:
      - description: TON wallet address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PlayerActivitySummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
      summary: Get queue summary for a wallet
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  TelegramAuth:
    description: Telegram Mini App authentication via initData string
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"